+++ src/js/item/baseItem.js
@@ -1,26 +1,46 @@
-import uniqid from 'uniqid';
+import DomHelper from "../helper/domHelper";
+import Uniqid from "uniqid";
 
 export default class BaseItem {
-    constructor(DomHelper, parameters) {
-        this.domHelper = DomHelper;
+    /**
+     * @param {object} parameters
+     */
+    constructor(parameters) {
+        this.domHelper = new DomHelper();
+        this.identifier = Uniqid();
         this.position = parameters.position;
-        this.backgroundColor = parameters.backgroundColor;
-        this.fontColor = parameters.fontColor;
-        this.title = parameters.title;
-        this.identifier = uniqid(parameters.title + '-');
     }
 
-    createIcon() {
-        let iconElement = this.domHelper.createElement('div', 'icon-button icon-radio-checked');
-        iconElement.setAttribute('data-for', this.identifier);
-        iconElement.style.left = this.position.left + 'px';
-        iconElement.style.top = this.position.top + 'px';
+    /**
+     * @returns {HTMLElement}
+     */
+    createHotspotElement() {
+        let element = this.domHelper.createElement('div', 'hotspot icon-radio-checked');
+        element.setAttribute('data-for', this.identifier);
+        element.style.left = this.position.left + 'px';
+        element.style.top = this.position.top + 'px';
 
-        return iconElement;
+        return element;
     }
 
-    createTitle() {
-        return this.domHelper.createElement('span', 'title', this.title);
+    /**
+     * @returns {HTMLElement}
+     */
+    createArrowElement() {
+        let element = this.domHelper.createElement('div', 'arrow-up');
+
+        return element;
+    }
+
+    /**
+     * @returns {HTMLElement}
+     */
+    createItemElement() {
+        let element = this.domHelper.createElement('div', 'item');
+        element.setAttribute('data-id', this.identifier);
+        element.append(this.createArrowElement());
+
+        return element;
     }
 
     renderHtml() {