+++ src/js/styles.js
@@ -1,3 +1,3 @@
 /* eslint-disable */
-import Styles from '../css/styles.css';
+import Styles from '../scss/main.scss';
 import Icomoon from '../css/icomoon.css';