+++ webpack.dev.js
@@ -1,0 +1,16 @@
+const merge = require('webpack-merge');
+const common = require('./webpack.common.js');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+
+module.exports = merge(common, {
+    output: {
+        filename: '[name].js'
+    },
+    plugins: [
+        new MiniCssExtractPlugin({
+            filename: '[name].css'
+        })
+    ],
+    devtool: 'source-map',
+    mode: 'development'
+});