+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.0.0
+* Added Picture items
+* Improved overall rendering
+* All images tags now have an alt attribute
+* The possibility to customize items styles in JS has been removed
+* Upgraded to Webpack 4
+* Now uses Sass as CSS preprocessor
+
+
 ## 1.0.0
 * First stable release.